// products/bosom-friend/project/bosom-friend-web/src/store/plugin/types/plat.type.ts
import type { PluginPlatformType } from '@web/store/plugin/types/baseTypes'

export interface XhsLoginStatus {
  home: boolean
  creator: boolean
}

export interface WxSphLoginStatus {
  channels: boolean
}

export interface PlatAccountInfo {
  type: PluginPlatformType
  loginCookie: string
  uid: string
  account: string
  avatar: string
  nickname: string
  fansCount?: number
  xhsLoginStatus?: XhsLoginStatus
  wxSphLoginStatus?: WxSphLoginStatus
}